refactor: Move paas/config to internal/cli/config
because after all that is the config of the cli
