+++ paas/client.go
@@ -23,12 +23,12 @@ import (
 	"github.com/spf13/pflag"
 	"github.com/suse/carrier/deployments"
 	"github.com/suse/carrier/internal/application"
+	"github.com/suse/carrier/internal/cli/config"
 	"github.com/suse/carrier/internal/duration"
 	"github.com/suse/carrier/internal/services"
 	"github.com/suse/carrier/kubernetes"
 	kubeconfig "github.com/suse/carrier/kubernetes/config"
 	"github.com/suse/carrier/kubernetes/tailer"
-	"github.com/suse/carrier/paas/config"
 	paasgitea "github.com/suse/carrier/paas/gitea"
 	"github.com/suse/carrier/paas/ui"
 	"k8s.io/apimachinery/pkg/labels"
